pyproject.toml
[tool.poetry]
name = "ai"
version = "0.1.0"
description = ""
authors = ["Arash Khazaei <0xclay@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.104.0"
uvicorn = "^0.23.2"
aiohttp = "3.9.0b0"
g4f = "^0.1.7.3"
sqlmodel = "^0.0.8"
alembic = "^1.12.0"
sqlalchemy = "1.4.40"
python-dotenv = "^1.0.0"
sqlalchemy-utils = "^0.41.1"


[tool.poetry.group.dev.dependencies]
black = "^23.10.0"
flake8 = "^6.1.0"
httpx = "^0.25.0"
pytest = "^7.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[flake8]
extend-ignore = "E203"
exclude = ".git,__pycache__,docs/source/conf.py,old,build,dist"
max-complexity = 10